+++ src/test/java/racingcar/game/RacingGameTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.CarDTO;
+import racingcar.domain.RacingGame;
 
 import java.util.List;
 import java.util.stream.Collectors;
